scripts and rules for combining nodes and edges
