+++ workflow/rules/join_network.smk
@@ -1,23 +1,22 @@
 # Take split .geoparquet files for nodes or edges and output unified .geoparquet
 # files for transport networks
 
-#   # do we care about joining nodes?
-#   rule join_network_nodes:
-#       input:
-#           lambda wildcards: expand(
-#               os.path.join("{OUTPUT_DIR}", "geoparquet", "{DATASET}_{FILTER_SLUG}", "slice-{i}_road_nodes_annotated.geoparquet"),
-#               **wildcards,
-#               i=range(config['slice_count'])
-#           ),
-#       output:
-#           "{OUTPUT_DIR}/{DATASET}_{FILTER_SLUG}/road_nodes.geoparquet",
-#       script:
-#           "../scripts/join_data.py"
-#
-#   """
-#   Test with:
-#   snakemake --cores all results/tanzania-mini_filter-highway-core/road_nodes.geoparquet
-#   """
+rule join_network_nodes:
+    input:
+        lambda wildcards: expand(
+            os.path.join("{OUTPUT_DIR}", "geoparquet", "{DATASET}_{FILTER_SLUG}", "slice-{i}_road_nodes_annotated.geoparquet"),
+            **wildcards,
+            i=range(config['slice_count'])
+        ),
+    output:
+        "{OUTPUT_DIR}/{DATASET}_{FILTER_SLUG}/road_nodes.geoparquet",
+    script:
+        "../scripts/join_data.py"
+
+"""
+Test with:
+snakemake --cores all results/tanzania-mini_filter-highway-core/road_nodes.geoparquet
+"""
 
 
 rule join_network_edges:
@@ -32,7 +31,7 @@ rule join_network_edges:
     output:
         "{OUTPUT_DIR}/{DATASET}_{FILTER_SLUG}/road_edges.geoparquet"
     script:
-        "../scripts/join_data.py"
+        "../scripts/join_edges.py"
 
 """
 Test with:
